Add files via upload
This was my first attempt at a large scale project. If you click the right buttons it works right. No error checking is involved, and many other glaring issues. I have since improved, but not that much. Enough to make fun of this at least. Would appreciate feedback if you can give that on this website.